# chore: speed up build by removing bzip and libz from critical path