--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -19,7 +19,7 @@ env:
 jobs:
   test:
     name: Test
-    runs-on: buildjet-8vcpu-ubuntu-2004
+    # runs-on: buildjet-8vcpu-ubuntu-2004
     timeout-minutes: 30
     steps:
       - name: Checkout Sources
@@ -106,7 +106,7 @@ jobs:
   calibnet-check:
     name: Calibnet sync check
     #if: ${{ false }}
-    runs-on: buildjet-8vcpu-ubuntu-2004
+    # runs-on: buildjet-8vcpu-ubuntu-2004
     timeout-minutes: 30
     steps:
       - name: Checkout Sources