--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -20,7 +20,8 @@ jobs:
   test:
     name: Test
     # runs-on: buildjet-8vcpu-ubuntu-2004
-    timeout-minutes: 30
+    # timeout-minutes: 30
+    runs-on: ubuntu-latest
     steps:
       - name: Checkout Sources
         uses: actions/checkout@v3
@@ -107,7 +108,8 @@ jobs:
     name: Calibnet sync check
     #if: ${{ false }}
     # runs-on: buildjet-8vcpu-ubuntu-2004
-    timeout-minutes: 30
+    # timeout-minutes: 30
+    runs-on: ubuntu-latest
     steps:
       - name: Checkout Sources
         uses: actions/checkout@v3