--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -19,9 +19,8 @@ env:
 jobs:
   test:
     name: Test
-    # runs-on: buildjet-8vcpu-ubuntu-2004
-    # timeout-minutes: 30
-    runs-on: ubuntu-latest
+    runs-on: buildjet-8vcpu-ubuntu-2004
+    timeout-minutes: 30
     steps:
       - name: Checkout Sources
         uses: actions/checkout@v3