--- a/node/db/Cargo.toml
+++ b/node/db/Cargo.toml
@@ -8,14 +8,14 @@ edition     = "2021"
 repository  = "https://github.com/ChainSafe/forest"
 
 [features]
-default = ["snappy", "lz4", "zlib", "bzip2"]
+default = ["lz4"]
 
-snappy = ["rocksdb/snappy"]
-lz4    = ["rocksdb/lz4"]
-zlib   = ["rocksdb/zlib"]
-bzip2  = ["rocksdb/bzip2"]
+lz4 = ["rocksdb/lz4"]
 # not included by default to reduce build time
 # uncomment when it needs to be used by other crates
+# snappy = ["rocksdb/snappy"]
+# zlib   = ["rocksdb/zlib"]
+# bzip2  = ["rocksdb/bzip2"]
 # zstd = ["rocksdb/zstd"]
 
 [dependencies]