--- a/node/forest_libp2p/Cargo.toml
+++ b/node/forest_libp2p/Cargo.toml
@@ -38,7 +38,6 @@ libp2p = { workspace = true, default-features = false, features = [
   "dns-async-std",
   "mplex",
   "request-response",
-  "websocket",
 ] }
 libp2p-bitswap.workspace = true
 log.workspace = true